Fixed NTSC video finally

+++ src/cpu.rs
@@ -2261,7 +2261,7 @@ mod tests {
             );
             assert_eq!(
                 cpu.bus.ppu.cycle_count,
-                ppu_cyc,
+                ppu_cyc as u32,
                 "ppu ${:02X} {:?} #{:?}",
                 instr.opcode(),
                 instr.op(),
